backend/requirements.txt
Django==3.2.20
djangorestframework==3.12.4
django-cors-headers==3.7.0
 psycopg2-binary==2.8.6